% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm.R
\name{gbm}
\alias{gbm}
\title{Geometric Brownian motion}
\arguments{
\item{n}{Number of samples to generate}

\item{s0}{Initial value}

\item{mu}{The mean-reverting value}

\item{sigma}{The annualized standard deviation of the noise term}
}
\description{
Geometric Brownian motion
}

