% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifs.R
\name{heighway}
\alias{heighway}
\title{Generate the Heighway Dragon}
\usage{
heighway(x)
}
\arguments{
\item{x}{A 2-by-n matrix representing a set of points}
}
\description{
This function is used to perform a single iteration of the
Heighway Dragon, which is simply the union of two functions
applied to the same set of points.
}
\examples{
library(lambda.tools)
# Create an instance with 12 iterations
x <- rbind(seq(0,1, by=0.05),0)
xn <- fold(1:12, function(a,b) heighway(b), x)
}
\author{
Brian Lee Yung Rowe
}
\keyword{math}

