% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraday.R
\name{intraday_ticks}
\alias{intraday_ticks}
\title{Create intraday ticks}
\usage{
intraday_ticks(period, hours)
}
\arguments{
\item{period}{In minutes}

\item{hours}{Hours of operation for exchange}
}
\value{
Vector of seconds used to offset a date. Compatible with POSIXct
}
\description{
Create intraday ticks
}

