% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractal.R
\name{next.seeds}
\alias{next.seeds}
\title{Create a new seed}
\usage{
next.seeds(old.seed, new.seed, pattern, idx, epoch)
}
\arguments{
\item{old.seed}{Previous seed used to generate pattern}

\item{new.seed}{Next seed used to generate pattern}

\item{pattern}{Available patterns to use}

\item{idx}{Index of current iteration}

\item{epoch}{Current epoch}
}
\description{
Create a new seed
}

