% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ou.R
\name{ou}
\alias{ou}
\title{Ornstein-Uhlenbeck process}
\arguments{
\item{n}{Number of samples to generate}

\item{s}{Initial value}

\item{lambda}{The shock, in annualized terms}

\item{mu}{The mean-reverting value}

\item{sigma}{The annualized standard deviation of the noise term}
}
\description{
Ornstein-Uhlenbeck process
}

