% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework.R
\name{plot_returns}
\alias{plot_returns}
\title{Plot asset prices and returns for fractal analysis}
\usage{
plot_returns(series, ...)
}
\arguments{
\item{series}{A time series}

\item{\dots}{Additional arguments to pass to plot}
}
\description{
This is a convenience function for studying the generated time series by the
fragtalrock package. Given a time series of prices, plotReturns will plot
both the original time series of prices and the returns series. This is a
useful visual aid in determining the utility of the simulated time series.
}

