% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifs.R
\name{polygon}
\alias{polygon}
\title{Construct a polygon}
\usage{
polygon(sides = 5, by = 0.1)
}
\arguments{
\item{sides}{The number of sides for the polygon}

\item{by}{The spacing between points}
}
\description{
Similar to \code{seq} but produces regular polygons instead of line
segments.
}
\examples{
p <- polygon(6)
}
\author{
Brian Lee Yung Rowe
}
\keyword{math}

