% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifs.R
\name{radians}
\alias{radians}
\title{Get radians from degrees}
\usage{
radians(degrees)
}
\arguments{
\item{degrees}{The angle in degrees}
}
\description{
Basic conversion from degrees to radians
}
\examples{
m <- radians(135)
}
\author{
Brian Lee Yung Rowe
}
\keyword{math}

