% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractal.R
\name{rfractal}
\alias{rfractal}
\title{Create time series based on fractal generators}
\arguments{
\item{n}{Number of samples to generate}

\item{initiator}{The initiator function. Defaults to \code{\link{rinitiator}}}

\item{generator}{The generator function. Defaults to \code{\link{rgenerator}}}
}
\value{
An xts object containing a time series of values representing asset prices
}
\description{
The fractal function generates a time series of points using basic principles
of fractal patterns. Fractal generation can be used to simulate a time series
of asset prices, which has been shown to better reflect the distribution of
returns than using a Gaussian random walk. Any number of points can be
generated based on specifying the total count or by running over a number of
epochs. The range of the data is defined by the given seed for the generation
plus the available patterns.
}
\section{Usage}{

rfractal(n, initiator=rinitiator, generator=rgenerator)
}
\references{
M. Frame, B. Mandelbrot, N. Neger. Fractal Geometry. 2009.
http://classes.yale.edu/fractals/
}

