% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractal.R
\name{rgenerator}
\alias{rgenerator}
\title{Create a generator pattern for fractal generation}
\arguments{
\item{xrange}{Range of x values}

\item{yrange}{Range of y values}
}
\description{
Create a generator pattern for fractal generation
}

