% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractal.R
\name{rinitiator}
\alias{rinitiator}
\title{Create an initiator pattern for fractal generation}
\arguments{
\item{xrange}{Range of x values}

\item{yrange}{Range of y values}
}
\description{
Create an initiator pattern for fractal generation
}

