% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifs.R
\name{rotation}
\alias{rotation}
\title{Create a 2x2 rotation matrix}
\usage{
rotation(degrees)
}
\arguments{
\item{degrees}{The number of degrees to rotate counter clockwise}
}
\description{
Create a rotation matrix for the specified number of degrees.
}
\examples{
m <- rotation(135)
}
\author{
Brian Lee Yung Rowe
}
\keyword{math}

