\name{sampleGenerators}
\docType{data}
\alias{sampleGenerators}
\title{Sample patterns for fractal generation of time series}
\description{ 
  This is a sample set of generators for creating time series using the 
  fragtalrock package.
}
\usage{sampleGenerators}
\format{ A list of available patterns to use as generators }
\source{ Inspiration }
\keyword{datasets}

