\name{sampleInitiators}
\docType{data}
\alias{sampleInitiators}
\title{Sample seed for fractal generation of time series}
\description{ 
  This is a sample seed aka intiator for creating time series using the 
  fractalrock package.

  In the future, multiple intiators may be included.
}
\usage{sampleInitiators}
\format{ A matrix representing xy coordinates for the seed }
\source{ Inspiration }
\keyword{datasets}

