% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifs.R
\name{sierpinski3}
\alias{sierpinski3}
\title{Generate the Sierpinski Gasket}
\usage{
sierpinski3(x)
}
\arguments{
\item{x}{A 2-by-n matrix representing a set of points}
}
\description{
This function generates a single iteration of the Sierpinski Gasket,
which replaces a triangle with 3 smaller triangles, one within each
vertex of the original triangle.
}
\examples{
x <- polygon(3)
xn <- fold(1:5, function(a,b) sierpinski3(b), x)
}
\author{
Brian Lee Yung Rowe
}
\seealso{
\code{\link{sierpinski4}} \code{\link{sierpinski5}}
}
\keyword{math}

