% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifs.R
\name{sierpinski5}
\alias{sierpinski5}
\title{Generate the Sierpinski Pentagon}
\usage{
sierpinski5(x)
}
\arguments{
\item{x}{A 2-by-n matrix representing a set of points}
}
\description{
This function generates a single iteration of the Sierpinski Pentagon,
which replaces a pentagon with 5 smaller pentagons where each vertex
contains a pentagon.
}
\examples{
library(lambda.tools)
x <- polygon(5)
xn <- fold(1:5, function(a,b) sierpinski5(b), x)
}
\author{
Brian Lee Yung Rowe
}
\seealso{
\code{\link{sierpinski3}} \code{\link{sierpinski4}}
}
\keyword{math}

