% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifs.R
\name{tile}
\alias{tile}
\title{Create random tiles}
\usage{
tile(x = c(0, 0), segments = 2, pop = c(0, 90, 180, 270))
}
\arguments{
\item{x}{Initial starting point}

\item{segments}{Number of segments to create}

\item{pop}{Population to sample angles from}
}
\description{
Create random tiles
}
\examples{
library(lambda.tools)
xs <- fold(1:4, function(a,b) tile(b), c(0,0))
xs <- fold(1:4, function(a,b) tile(b), polygon(20))
}

