% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework.R
\name{trading_dates}
\alias{trading_dates}
\title{Generate trading dates with a given calendar}
\arguments{
\item{period}{The date period, defaulting to 1 day (1440 minutes).
Units are in minutes, so values less than 1440 will result in 
intraday time steps being added.}
}
\description{
Generate trading dates with a given calendar
}
\examples{
trading_dates('2009-02-24',obs=10)

th <- function(x) trading_hours(x,'cme')
trading_dates('2014-01-01','2014-03-31', 5, th)
}

