% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm.R
\name{wiener}
\alias{wiener}
\title{Wiener process}
\arguments{
\item{n}{Number of samples to generate}

\item{s}{Initial value}

\item{mu}{The mean-reverting value}

\item{sigma}{The annualized standard deviation of the noise term}
}
\description{
Wiener process
}

