\name{sampleInitiator}
\docType{data}
\alias{sampleInitiator}
\title{Sample seed for fractal generation of time series}
\description{ 
  This is a sample seed aka intiator for creating time series using the 
  fragtalrock package.

  In the future, multiple intiators may be included.
}
\usage{sampleInitiator}
\format{ A matrix representing xy coordinates for the seed }
\source{ Inspiration }
\keyword{datasets}

